Squashed commit of the following:
commit 65ef93d8e11f03b2d825f40b41a3bccb821abf07
Merge: 08017cb 54e6080
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Tue Aug 25 14:58:34 2009 +0200

    Merge branch 'master' into urho-sync

    Conflicts:
    	tests/libtracker-data/tracker-ontology-test.c
    	tests/libtracker-data/tracker-sparql-test.c

commit 08017cbb8a3e1a373f88c6842fb526e916009025
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Tue Aug 25 12:45:56 2009 +0200

    Limiting the size of the log file

commit 2b6623eae4c52c232d05ae01ab77bdaca0f9d1ba
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 16:29:11 2009 +0200

    Added some code comments

commit 28e0498331c6fe0914528cd17f035339931bfa53
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 16:14:44 2009 +0200

    Start save of backup over DBus after queue's commit

commit 4c59007564f1138c01a66e2e45f891d157da03e8
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 15:57:19 2009 +0200

    Fixed restore on startup of meta.db is missing and meta-backup.db isn't

commit e43ce2eaadda7759c3e5ca7aa73c8e40543343e2
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 15:17:37 2009 +0200

    Adding kill to soft-reset

commit b7866eded285c138673c767e2979342c89fa1b37
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 14:50:38 2009 +0200

    Made DBus's Restore() synchronous

commit 70eae9cb2c4747d226e687e10452e6e4054fcd27
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 14:23:31 2009 +0200

    Fixed path of the meta-backup.db file

commit bbc39e88fa6affbe570676a304b7eee520e7e8e1
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 14:21:25 2009 +0200

    Fixed restore after tracker-processes -s

commit a98ac7a5fcc9c4a6957ba292274a2211913ce06d
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 13:31:08 2009 +0200

    Revert changes to INSTALL file

commit 0016d9d9a0811abb9a098b49da64db6c1febee86
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 12:39:22 2009 +0200

    Code cleanup

commit a10c615c63e35c548d3255cdc93632012ab5ae69
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 12:24:02 2009 +0200

    Code cleanup

commit 3de8d7d0a2447840bd4c4168a9993fc78d504ad3
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 12:02:58 2009 +0200

    Added transaction barriers to the journal player

commit 6af3250b97b79c4adefa518306081c1e86e42be3
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 12:00:15 2009 +0200

    Improvement for tracker-processes util

commit 636d9670bd79ca4f86350acd85818fcd5679527e
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 11:47:16 2009 +0200

    Fixed finding location of meta-backup.db

commit d487838385db6ec4b3f7d89fc11440be60813a73
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 24 11:44:40 2009 +0200

    Play journal always when meta.db didn't exist

commit 013b277d1f3079d228f20a8fc14b11da8d5f8f6e
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Fri Aug 21 17:11:44 2009 +0200

    Changed the format of the log, playing it using mmap

commit a61dbf091354411c093674105039e1546e7a3151
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Fri Aug 21 11:19:23 2009 +0200

    No need to do an extra fsync of commit causes one already

commit e329dbca27a1ee5258757a0ff0c051097ecb93a6
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Fri Aug 21 11:12:39 2009 +0200

    Fixed logic, the timeout is for backup, the max for fsync

commit ed462a8126c9a2314d4adb516183340067a96b7b
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Fri Aug 21 11:06:14 2009 +0200

    Fixed another glitch

commit 67fa0312d568718effb797d77308b5133d3d76e2
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Fri Aug 21 11:00:46 2009 +0200

    Make it exact

commit d5ab45435cc1a2e9b56fbafb04a82f150c843e51
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Fri Aug 21 10:58:04 2009 +0200

    Make fsync happen less often on the journal

commit 250362370069b56cd0b134fb7db02d9b561c52ef
Author: Jürg Billeter <j@bitron.ch>
Date:   Fri Aug 21 09:37:35 2009 +0200

    Remove raptor.h include

commit 18048eeda22eb94030a32e6f59e4c5ca0fa53579
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Thu Aug 20 15:35:30 2009 +0200

    Several bugfixes in backup and restore

commit 0bdb823b6b13081ff88a98c0e9bc0bb62d655e24
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Thu Aug 20 15:09:38 2009 +0200

    Playing the journal on restore

commit b937a201f7588a4f401bc13f51e47a4754e1e3b2
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Thu Aug 20 14:58:54 2009 +0200

    Changed the paths of the meta-backup and the journal

commit 142a32f01e4a43275d954e48b1134ba4a7a965db
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Thu Aug 20 12:47:04 2009 +0200

    Backup and restore support

commit 2ef24a4563c2e6fab6f17b0a1926602135639cb4
Merge: a7a7339 8a66865
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Thu Aug 20 10:56:00 2009 +0200

    Merge branch 'master' into urho-sync

commit a7a73398baa8d87ecdb32b0228ffb78c5818a36e
Merge: 2c25a57 d8aebcb
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Wed Aug 19 15:39:40 2009 +0200

    Merge branch 'master' into urho-sync

    Conflicts:
    	src/tracker-store/tracker-main.c

commit 2c25a572dbadf53ed5d9316d1eb61c0655335e98
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Wed Aug 19 15:30:58 2009 +0200

    Atomic rename and journal load

commit a3eec4151daadc9c575edc0174e2e610e50e61e4
Merge: 82003a5 377dc98
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Wed Aug 19 15:16:18 2009 +0200

    Merge branch 'master' into urho-sync

commit 82003a5114bb6fd3ee69190be56e251afb7adace
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Wed Aug 5 11:16:47 2009 +0200

    Don't perform the check when need_journal isn't passed

commit 57e8168e31fcd5322798bd2b66f1f079b100ed76
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Wed Aug 5 10:43:02 2009 +0200

    Refactored the journal code a bit

commit b37be6edbf55d3b43960fa87fd3b0322c9ae121a
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Tue Aug 4 16:42:39 2009 +0200

    Fixed a critical warning

commit 06fee84f767a2363c8eac30ccd424ac1f932d975
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Tue Aug 4 16:35:55 2009 +0200

    Changed the logic about journal playing and meta-backup.db picking

commit fab94fb68cc9b044ac84641ab4f80fba387233fd
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Tue Aug 4 14:11:21 2009 +0200

    Comment with a warning about the journal in a rare case

commit 94ae1fbfa25b6c2faa1e073ec171c5f5bc6cb3d6
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Mon Aug 3 18:28:36 2009 +0200

    Restoring the db using a backup on corruption

commit 47e8885909a6b89559788a1f789d2902eedadfb2
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Fri Jul 31 17:45:19 2009 +0200

    Setting sync=off as the journal is being made now

commit f697a726b69df2e91b5c5daeb31e9252fa2a3ba2
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Fri Jul 31 17:43:28 2009 +0200

    Added missing files

commit 9bcfa4e19b7042c7c315eecb9fb7d7939050bf36
Author: Philip Van Hoof <philip@codeminded.be>
Date:   Fri Jul 31 17:41:43 2009 +0200

    Experiment for journal synchronization mode

+++ src/libtracker-db/tracker-db-manager.c
@@ -35,8 +35,10 @@
 #include <libtracker-common/tracker-type-utils.h>
 #include <libtracker-common/tracker-utils.h>
 
+#include "tracker-db-backup.h"
 #include "tracker-db-manager.h"
 #include "tracker-db-interface-sqlite.h"
+#include "tracker-db-interface.h"
 
 /* ZLib buffer settings */
 #define ZLIB_BUF_SIZE		      8192
@@ -152,6 +154,8 @@ static gboolean		   db_exec_no_reply    (TrackerDBInterface *iface,
 						const gchar	   *query,
 						...);
 static TrackerDBInterface *db_interface_create (TrackerDB	    db);
+static TrackerDBInterface *tracker_db_manager_get_db_interfaces     (gint num, ...);
+static TrackerDBInterface *tracker_db_manager_get_db_interfaces_ro  (gint num, ...);
 
 static gboolean		   initialized;
 static gchar		  *sql_dir;
@@ -656,7 +660,7 @@ db_set_params (TrackerDBInterface *iface,
 	       gint		   page_size,
 	       gboolean		   add_functions)
 {
-	tracker_db_interface_execute_query (iface, NULL, "PRAGMA synchronous = NORMAL;");
+	tracker_db_interface_execute_query (iface, NULL, "PRAGMA synchronous = OFF;");
 	tracker_db_interface_execute_query (iface, NULL, "PRAGMA count_changes = 0;");
 	tracker_db_interface_execute_query (iface, NULL, "PRAGMA temp_store = FILE;");
 	tracker_db_interface_execute_query (iface, NULL, "PRAGMA encoding = \"UTF-8\"");
@@ -861,7 +865,7 @@ db_interface_create (TrackerDB db)
 }
 
 static void
-db_manager_remove_all (void)
+db_manager_remove_all (gboolean rm_backup_and_log, gboolean not_meta)
 {
 	guint i;
 
@@ -871,9 +875,33 @@ db_manager_remove_all (void)
 	 * calculate the absolute directories here. 
 	 */
 	for (i = 1; i < G_N_ELEMENTS (dbs); i++) {
+
+		if (not_meta && i == TRACKER_DB_METADATA) {
+			continue;
+		}
+
 		g_message ("  Removing database:'%s'",
 			   dbs[i].abs_filename);
 		g_unlink (dbs[i].abs_filename);
+	}
+
+	if (rm_backup_and_log) {
+		GFile *file;
+		gchar *path;
+
+		file = tracker_db_backup_file (NULL, TRACKER_DB_BACKUP_META_FILENAME);
+		path = g_file_get_path (file);
+		g_message ("  Removing database:'%s'",
+			   path);
+		g_free (path);
+		g_file_delete (file, NULL, NULL);
+		g_object_unref (file);
+		path = tracker_db_journal_filename ();
+		g_message ("  Removing database:'%s'",
+			   path);
+		file = g_file_new_for_path (path);
+		g_file_delete (file, NULL, NULL);
+		g_object_unref (file);
 	}
 }
 
@@ -1016,17 +1044,78 @@ tracker_db_manager_ensure_locale (void)
 	g_free (stored_locale);
 }
 
+static gboolean
+check_meta_backup (gboolean *did_copy)
+{
+	const gchar *meta_filename;
+	gboolean retval = FALSE;
+
+	/* This is currently the only test for need_journal. We should add a
+	 * couple tests that test meta.db against consistenty, and if not
+	 * good, copy meta-backup.db over and set need_journal (being less 
+	 * conservative about using the backup, and not trusting the meta.db
+	 * as much as we do right now) */
+
+	meta_filename = dbs[TRACKER_DB_METADATA].abs_filename;
+
+	if (meta_filename) {
+		GFile *file;
+
+		file = g_file_new_for_path (meta_filename);
+
+		/* (more) Checks for a healthy meta.db should happen here */
+
+		if (!g_file_query_exists (file, NULL)) {
+			GFile *backup;
+
+			backup = tracker_db_backup_file (NULL, TRACKER_DB_BACKUP_META_FILENAME);
+
+			if (g_file_query_exists (backup, NULL)) {
+				GError *error = NULL;
+
+				/* Note that we leave meta-backup.db as is, it'll
+				 * be overwritten first-next time tracker-store.c's
+				 * sync_idle_handler will instruct this. */
+
+				g_file_copy (backup, file, G_FILE_COPY_OVERWRITE,
+				             NULL, NULL, NULL, &error);
+
+				if (!error && did_copy) {
+					*did_copy = TRUE;
+				}
+
+				g_clear_error (&error);
+			}
+
+			/* We always play the journal in case meta.db wasn't
+			 * healthy. Also if meta-backup.db didn't exist: that
+			 * just means that tracker-store.c's sync_idle_handler
+			 * didn't yet ran (meanwhile a first log-file is yet
+			 * already being made) */
+
+			retval = TRUE;
+
+			g_object_unref (backup);
+		}
+
+		g_object_unref (file);
+	}
+
+	return retval;
+}
+
 gboolean
 tracker_db_manager_init (TrackerDBManagerFlags	flags,
 			 gboolean	       *first_time,
-			 gboolean	        shared_cache)
+			 gboolean	        shared_cache,
+			 gboolean	       *need_journal)
 {
 	GType		    etype;
 	TrackerDBVersion    version;
 	gchar		   *filename;
 	const gchar	   *dir;
 	const gchar        *env_path;
-	gboolean	    need_reindex;
+	gboolean	    need_reindex, did_copy = FALSE;
 	guint		    i;
 
 	if (first_time) {
@@ -1126,6 +1215,13 @@ tracker_db_manager_init (TrackerDBManagerFlags	flags,
 		}
 	}
 
+	if (need_journal) {
+		/* That did_copy is used for called db_manager_remove_all, we
+		 * don't want it to also remove our freshly copied meta.db file. */
+
+		*need_journal = check_meta_backup (&did_copy);
+	}
+
 	/* If we are just initializing to remove the databases,
 	 * return here. 
 	 */
@@ -1158,11 +1254,15 @@ tracker_db_manager_init (TrackerDBManagerFlags	flags,
 		 * will cause errors and do nothing.
 		 */
 		g_message ("Cleaning up database files for reindex");
-		db_manager_remove_all ();
+
+		/* Remove all but the meta.db in case of did_copy, else remove
+		 * all (only in case meta-backup.db was not restored) */
+
+		db_manager_remove_all (FALSE, did_copy);
 
 		/* In cases where we re-init this module, make sure
 		 * we have cleaned up the ontology before we load all
-		 * new databases.
+		* new databases.
 		 */
 		tracker_ontology_shutdown ();
 
@@ -1214,7 +1314,56 @@ tracker_db_manager_init (TrackerDBManagerFlags	flags,
 								    TRACKER_DB_CONTENTS,
 								    TRACKER_DB_COMMON);
 	}
+
 	return TRUE;
+}
+
+void 
+tracker_db_manager_disconnect (void)
+{
+	if (resources_iface) {
+		guint i;
+ 		TrackerDB attachments[3] = { TRACKER_DB_FULLTEXT,
+					     TRACKER_DB_CONTENTS,
+					     TRACKER_DB_COMMON };
+
+		for (i = 0; i < 3; i++) {
+			TrackerDB db = attachments [i];
+
+			db_exec_no_reply (resources_iface,
+					  "DETACH '%s'",
+					  dbs[db].name);
+		}
+
+		tracker_db_interface_disconnect (resources_iface);
+	}
+}
+
+void 
+tracker_db_manager_reconnect (void)
+{
+	if (resources_iface) {
+		guint i;
+ 		TrackerDB attachments[3] = { TRACKER_DB_FULLTEXT,
+					     TRACKER_DB_CONTENTS,
+					     TRACKER_DB_COMMON };
+
+		tracker_db_interface_reconnect (resources_iface);
+
+		db_set_params (resources_iface,
+			       dbs[TRACKER_DB_METADATA].cache_size,
+			       dbs[TRACKER_DB_METADATA].page_size,
+			       TRUE);
+
+		for (i = 0; i < 3; i++) {
+			TrackerDB db = attachments [i];
+
+			db_exec_no_reply (resources_iface,
+					  "ATTACH '%s' as '%s'",
+					  dbs[db].abs_filename,
+					  dbs[db].name);
+		}
+	}
 }
 
 void
@@ -1268,11 +1417,11 @@ tracker_db_manager_shutdown (void)
 }
 
 void
-tracker_db_manager_remove_all (void)
+tracker_db_manager_remove_all (gboolean rm_backup_and_log)
 {
 	g_return_if_fail (initialized != FALSE);
 	
-	db_manager_remove_all ();
+	db_manager_remove_all (rm_backup_and_log, FALSE);
 }
 
 void
@@ -1327,7 +1476,7 @@ tracker_db_manager_get_file (TrackerDB db)
  *
  * returns: (caller-owns): a database connection
  **/
-TrackerDBInterface *
+static TrackerDBInterface *
 tracker_db_manager_get_db_interfaces (gint num, ...)
 {
 	gint		    n_args;
@@ -1361,7 +1510,7 @@ tracker_db_manager_get_db_interfaces (gint num, ...)
 	return connection;
 }
 
-TrackerDBInterface *
+static TrackerDBInterface *
 tracker_db_manager_get_db_interfaces_ro (gint num, ...)
 {
 	gint		    n_args;
